requirements.txt
requests==2.9.1
pandas==0.20.2
persist_queue==0.3.3
numpy==1.12.1
scikit_learn==0.19.1
PyYAML==3.12
